+++ example/lib/main.dart
@@ -7,8 +7,9 @@ void main() {
   FlipperClient flipperClient = FlipperClient.getDefault();
 
   flipperClient.addPlugin(new FlipperNetworkPlugin());
-  flipperClient.addPlugin(new FlipperReduxInspectorPlugin());
   flipperClient.addPlugin(new FlipperSharedPreferencesPlugin());
+  // flipperClient.addPlugin(new FlipperDatabaseBrowserPlugin());
+  flipperClient.addPlugin(new FlipperReduxInspectorPlugin());
   flipperClient.start();
 
   runApp(MyApp());