+++ lib/plugins/plugins.dart
@@ -1,3 +1,4 @@
+export './database/flipper_dbbrowser_plugin.dart';
 export './network/flipper_network_plugin.dart';
 export './redux/flipper_reduxinspector_plugin.dart';
 export './sharedpreferences/flipper_shared_preferences_plugin.dart';